fix: fix media validator
