+++ src/mocks/mockMedias/mockMedia.type.ts
@@ -1,9 +1,9 @@
-export type LoadImage = {
+export type Image = {
   addEventListener: (event: string, cb: () => void) => void;
   src: string;
 };
 
-export type LoadVideo = {
+export type Video = {
   addEventListener: (event: string, cb: () => void) => void;
   duration: number;
   height: number;
