--- src/utils/checkMediaAndText/checkMedia/checkMedia.ts
@@ -1,3 +1,0 @@
-export function checkMediaSize(file: File, limitSize: number): boolean {
-  return file.size <= limitSize;
-}
