--- src/utils/checkMediaAndText/checkText/checkText.ts
@@ -1,3 +1,0 @@
-export function checkTextLength(text: string, limitLength: number): boolean {
-  return text.length <= limitLength;
-}
