+++ src/utils/mediaAndTextValidator/mediaValidator/imageValidator/imageData.type.ts
@@ -1,0 +1,4 @@
+export type ImageData = {
+  height: number;
+  width: number;
+};
