+++ src/utils/mediaAndTextValidator/mediaValidator/interfaces/fileValidator.ts
@@ -1,0 +1,5 @@
+export type FileValidator = {
+  aspectRatio: (limitAspectRatio: string) => Promise<boolean>;
+  checkSize: (limitSize: number) => boolean;
+  resolution: (limitWidth: number, limitHeight: number) => Promise<boolean>;
+};
