+++ src/utils/mediaAndTextValidator/mediaValidator/mediaValidator.spec.ts
@@ -1,0 +1,15 @@
+import { mockFileImage, mockFileVideo } from '~mocks/mockMedias/mockFiles';
+
+import { MediaValidator } from './mediaValidator';
+
+const image = new MediaValidator(mockFileImage);
+const video = new MediaValidator(mockFileVideo);
+
+describe('check medias', () => {
+  it('limit media size', () => {
+    const limitSize = 9999;
+
+    expect(image.checkSize(limitSize)).toBe(true);
+    expect(video.checkSize(limitSize)).toBe(true);
+  });
+});
