+++ src/utils/mediaAndTextValidator/mediaValidator/videoValidator/videoData.type.ts
@@ -1,8 +1,3 @@
-export type ImageData = {
-  height: number;
-  width: number;
-};
-
 export type VideoData = {
   duration: number;
   height: number;
